Limit how much data is sent/received over a TCP socket at once.
When using `twib coredump`, at least with commercial games, the core
dump is around 3GB.  twibd gathers it all in memory and then tries to
send it all at once to twib.  However, on my macOS system, send() and
recv() produce EINVAL when called with such large size arguments.
To work around this, change SocketMessageConnection to limit the amount
of data transferred in a single send/recv call to 1MB.

With this change, core dumping works on my system but uses an enormous
amount of RAM: twib peaks at 18GB of RAM usage.  So this change, while
necessary for correctness, is not sufficient to make core dumping work
nicely. :)